requirements.txt
black==25.12.0
isort==7.0.0
loguru==0.7.3
pytest==9.0.2
ruff==0.14.10
selenium==4.39.0
pytest-mock==3.15.1
jsonschema==4.25.1
requests==2.32.5
pyyaml==6.0.3
